Refactor plugin into BotService
