+++ src/main/kotlin/cancelcloud/command/AutoUpdatesCommand.kt
@@ -1,18 +1,17 @@
 package cancelcloud.command
 
-import cancelcloud.PurpurInsightPlugin
+import cancelcloud.service.BotService
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent
 
 object AutoUpdatesCommand {
     operator fun invoke(event: SlashCommandInteractionEvent) {
         val minutes = event.getOption("minutes")?.asLong
-        val plugin = PurpurInsightPlugin.instance
         if (minutes == null) {
-            event.reply("Aktuelles Intervall: ${plugin.autoUpdateMinutes} Minuten")
+            event.reply("Aktuelles Intervall: ${BotService.intervalMinutes} Minuten")
                 .setEphemeral(true).queue()
             return
         }
-        plugin.updateInterval(minutes)
+        BotService.updateInterval(minutes)
         if (minutes <= 0) {
             event.reply("Automatische Updates deaktiviert.").queue()
         } else {
